+++ src/DspProgressionStatusExporter/CompactSnapshotBuilder.cs
@@ -38,7 +38,7 @@ namespace DspProgressionStatusExporter
                     2301, 2302, 5002, 6003
                 } },
                 { "yellow", new int[] { 6003 } },
-                { "purple", new int[] { 6004 } },
+                { "purple", new int[] { 6004, 1303, 1124, 1402 } },
                 { "green", new int[] { 6005, 1305, 1209 } },
                 { "dyson", new int[] { 1501 } },
                 { "photon", new int[] { 1208, 1122 } },